JavaScriptCompiler: fix debug output
In the calculated endianess branch, the compiler was changed to always
emit the _read function instead of only when debug was enabled. The
readHeader method was introduced to emit the header of this function.
The JavaScriptCompiler still had code in the classConstructorHeader
method to emit the _read header if debug.