Implement Go-specific switch handling.
Go does not support switching on slices except for to compare with nil.

In order to handle byte slice switches in Go, we can use the switch on
true technique and the bytes.Equal function.

To facilitate this, a new Go-specific component called GoSwitchOps is
introduced, which mainly facilitates use of switch-on-true + comparison
functions. Unfortunately, this will need to be special-cased whenever
other slice types are needed. For truly general switches,
reflect.DeepEqual could be used, but is too slow to use in the common
cases.