Fix writing enums in repetitions
When you want to write enum value, you need to call method
translator.enumToInt, which expects its arguments of type Ast.expr.
So we need a way to convert `Identifier` to `Ast.expr.Name`. I thought
that method Identifier.toAstIdentifier would serve this purpose,
but it doesn't work and I don't know how it was supposed to work.
Anyway, the conversion had been done as
`Ast.expr.Name(Ast.identifier(idToStr(id)))`. However,
this simple conversion is wrong. The `Identifier` is language-agnostic
object, but idToStr(id) converts it to language-dependent name.

So when you try to call `translator.translate(
  Ast.expr.Subscript(
    Ast.expr.Name(Ast.identifier(idToStr(id))),
    Ast.expr.Name(Ast.identifier(Identifier.INDEX))
  )
)` with id = NamedIdentifier('some_array'), idToStr(id) will return
'someArray' for Java, and the execution will fail when detecting
type of the container you try to subscript, because there is no field
with name 'someArray' (actualy, the ClassTypeProvider.determineType
method internally creates NamedIdentifier(attrName), so it fails
on constraint ^[a-z][a-z0-9_]*$).

Hence I felt the need to inject language-specific code to AST object,
so I created class Ast.expr.CodeLiteral(code) for this purpose.
The BaseTranslator for it just returns the `code` itself. In the future,
we may need also specify the DataType of the code literal, so that
detecting type on it doesn't fail, but I haven't needed it yet.