Don't prepend "src" to output dirs of Java and Go.
Java and Got seem to be the only languages prepending "src" to their output directories, which make them fail in case a Maven-like directory tree with "src/main/java" is used. "java" itself needs to be the source folder already, not "java/src". Additonally its easy for users to add paths on their own, but difficult to remove those added by KS.

Like has been disucssed in the correpsonding issue, "src" keeps getting added in case "--target all" is given, which is most likely only the case for tests and those use a dir structure of "java/bin|src", where that makes sense.

https://github.com/kaitai-io/kaitai_struct/issues/287