Ignore and remove IntelliJ IDEA project files
These files shouldn't be in the repo. Most of them are automatically
generated by the IDEA Scala plugin when importing the sbt project, so
tracking them in the repo is unnecessary and annoying (the plugin
sometimes regenerates the IDEA files even if the sbt project hasn't
changed, which leads to unnecessary changes/commits).
The few non-auto-generated IDEA project files are either personal
settings (e. g. project name) or factory defaults, which don't belong
in the repo either.

The only downside of this change is that pulling it for the first time
will break existing local IDEA projects. Not much can be done about
this - though as a workaround you can make a backup copy of your .idea
directory before pulling, and after pulling close the IDEA project,
delete the broken .idea directory and replace it with the backup copy.