Fix substring() in Lua
Lua indexes strings and arrays from 1 instead of 0, and the
string.sub(from, to) function includes both characters at
offset `from` and `to`. So the `to` parameter stays as it is
(indexing from 1 is compensated by being inclusive instead
of exclusive), but `from` parameter needs to be increased
by 1 (as index 0 in KS language has equivalent 1 in Lua).