Lua, PHP: avoid conversion of "unsigned" 64-bit int literals to float
PHP and Lua only have *signed* 64-bit integers, so a decimal literal
greater than the maximum signed 64-bit integer (`2**63 - 1`) will be
interpreted as float. This is usually unacceptable, because it loses
precision. If we output integers `2**63 <= x <= 2**64 - 1` so that they
are interpreted as signed integers, we lose the information that they
were unsigned, but at least we don't lose any significant bits.

This approach is consistent with how integer reading methods in runtime
libraries work - read unsigned 64-bit ints as signed values if the
language doesn't support the unsigned.