Use translator.doIntLiteral in enum defs, not naive int->str conv
Excluding GoCompiler, because all enums are currently typed as `int` but
this is unknown to the doIntLiteral function, which generates `int64`,
`uint32`, etc. This prevents the generated Go code from compiling,
because types with explicit size (`int64`, `uint32`, ...) cannot be
assigned to `int`.

NimCompiler is also excluded, because I haven't adapted
NimTranslator.doIntLiteral yet.