Check Python runtime version using tuple comparison (#208)
The previous implementation used pkg_resources.parse_version to parse
the kaitaistruct module's __version__ string. Unfortunately, importing
pkg_resources is relatively slow, which had a noticeable performance
impact on every ksc-generated parser.

The new implementation uses the newly added kaitaistruct.API_VERSION
attribute, which stores the runtime version as a tuple of ints. In this
format the version information can be compared directly without having
to parse it from a string first. This allows removing the import of
pkg_resources, which fixes the performance problem.

Part of the fix for kaitai-io/kaitai_struct#804.