ErrorMessagesSpec: enable inserting empty lines in expected output
After changing the format of problem listing in 977df8fa, each problem
is now separated by an empty line from the rest of the output. To fix
ErrorMessagesSpec tests, the expected outputs in formats_err/*.ksy specs
need to be updated, but the problem is how to insert empty lines. If
each line of the expected output in the .ksy spec is required to start
with '# ', the empty line would have to be exactly '# ', but that relies
on the trailing space and I have my VS Code configured to *trim*
trailing whitespace. I don't plan to change this setting, and requiring
the trailing space would be a stupid and extremely confusing thing to
do, because it's invisible (good luck with detecting its presence in the
console output - you'd probably have to display the contents as a hex
dump).

So I relaxed the expected output line detection so that it can start
with just '#' (no longer requiring '# '). An additional space may (in
which case it's stripped) or may not be present.