Explain the issue with `-d` on Unix-like shells in the usage text
See https://github.com/kaitai-io/kaitai_struct/issues/289#issuecomment-1018561403

I know that it actually doesn't matter if the shell is Unix-like or not,
but rather if the shell wrapper (and not the .bat script launcher) is
being used to launch `kaitai-struct-compiler`. However, I think the
classification I chose is more understandable by users, even though it's
slightly incorrect - I guess it'll work most of the time.