_write(): use special loops for each repetition type again
Using a common loop (see
https://github.com/kaitai-io/kaitai_struct_compiler/pull/182) would be
fine if evaluating the user expressions in `repeat-expr` and
`repeat-until` couldn't cause any side effects. However, it in fact
*can* cause them by referring to positional instances, which need to be
written at the moment they're first invoked so that the context of all
`_io` properties matches the context in which they would be parsed.

Hence the first invocation of instances has a side effect, and not
invoking appropriate instances by not evaluating an expression (used in
parsing) in writing would therefore introduce a discrepancy between
parsing and serialization behavior.

In general, if a user expression we know nothing about is used somewhere
in parsing, we must use it in the same place during serialization.