Move consistency checks from _check() to _write() if needed
Also using a common loop for all kinds of repetitions once again
(previously, the common loop was used, but in
https://github.com/kaitai-io/kaitai_struct_compiler/commit/98978181af5e034690e8d881acf97f270ba9ad0e
the loops were made specific for each repetition type). If evaluating a
`repeat-expr` or `repeat-until` expression can cause side effects (i.e.
when it refers to a parse instance), consistency checks associated with
these expressions will be performed in _write() and they will invoke the
appropriate parse instances. This ensures that the behavior will be the
same as in _read() while the common "foreach" loop is used.