Remove faulty Identifier.toAstIdentifier() method
Contrary to what the docstring says ("Identifier ready to be used in KS
expressions."), nothing could be further from the truth since it relies
on `humanReadable`, which is (as the name suggests) merely a
*human-readable* identifier, i.e. completely unusable for the purpose
stated here.

For this reason, using the `toAstIdentifier()` method in a hope that it
would work as declared often led to bugs, because for example it didn't
work for unnamed fields (`NumberedIdentifier`), as they normally cannot
appear in user-supplied expressions. Fortunately, all former usages of
`toAstIdentifier()` were already changed in commit e535b209 to use
`Ast.expr.InternalName`, which was specifically designed for the purpose
of injecting any identifiers we need (including unnamed fields) when
crafting an AST internally in the compiler. So we can finally remove
`toAstIdentifier()` for good because it is no longer used.