Stop using `_io` when throwing validation errors from _check()
Relevant tests were added in these commits:

- Java: https://github.com/kaitai-io/kaitai_struct_tests/commit/e92fb336fb2eda17521224ab05ea416f8052828e
- Python: https://github.com/kaitai-io/kaitai_struct_tests/commit/e7869f0ce9a1bd9dc58f5523366615fe9e1fbbc8

This commit is needed for the `testCheckBadValidOldIo` /
`test_check_bad_valid_old_io` test methods to pass.

The _check() method is intended to verify pure data consistency and is
supposed to be called at the time when the actual `_io` is not available
yet (or is not in the correct state) and should not be used even if it's
not `null`. Before this commit, if we wanted to initialize a KS object
by reading an existing stream and then edit the data and write them,
_check() would read the position from the old `_io` used for reading and
report it in the validation error, which is wrong.