--no-auto-read: save user objects to array before their _read() fails
This allows visualizers like the Web IDE to show more (partial) output
in case of a parsing error - this feature was suggested in
https://github.com/kaitai-io/kaitai_struct_webide/issues/156.

The previous order of operations (creating a new user type object and
saving it to a local temporary variable `_t_*`, calling `_t_*._read()`
and then pushing `_t_*` into the array) means that if _read() fails, the
partial object is not accessible from the outside world, because it was
only stored in the local temporary variable. Pushing the object to the
public array before the _read() call makes the partial object accessible
even in case _read() fails.