Generate imports also for parameters and arrays
This partially fixes the ImportsParamsDefUsertypeImported and
ImportsParamsDefArrayUsertypeImported tests for C++ (but not quite -
they both remain `failed` due to other issues)

Technically, importing a module just because a type from it is accepted
via a parameter is not needed in dynamically typed languages without any
type annotations (which means that this commit adds unnecessary imports
in JavaScript, Lua, Perl and Python), but that's just the way it works
right now. `TypeProcessor` cannot distinguish which languages need to
import the spec in this specific case and which don't (that ultimately
depends on the particular language implementation).

As far as I know, unnecessary imports don't really cause any problems,
so that's fine by me. If we wanted to avoid them, we could remove
`TypeProcessor` altogether and only add the imports in language-specific
`*Compiler` classes (right next to places where it is referenced
somehow, e.g. in a parameter declaration, type hint, etc.), but let's
leave that for the future.