C++: forward-declare the local class instead of external types
Fixes tests EnumToIClassBorder1 and ImportsParamsDefUsertypeImported
for C++

If we always put a forward declaration at the beginning of the .h header
file (even before the section with `#include`), then we don't have to
worry about needing forward declarations for external types in the case
of circular includes. These were needed before if a header file A (the
main one with the class that will be called from the outside) included a
header file B and vice versa, for example (see
[imports_circular_a.ksy](https://github.com/kaitai-io/kaitai_struct_tests/blob/ec0fafac3143741ced11ddda3e296efc2e9822bb/formats/imports_circular_a.ksy)
and [imports_circular_b.ksy
](https://github.com/kaitai-io/kaitai_struct_tests/blob/ec0fafac3143741ced11ddda3e296efc2e9822bb/formats/imports_circular_b.ksy)).
Due to header guards, the currently processed header file is not
processed a second time. So the header file A starts being processed,
but it has an `#include` of the header file B very early on. So the
processing moves to the header file B, which has an `#include` of the
header file A, but it is ignored because the header file A is already
marked as included. Since the definition of the class in the header file
B depends on the class from the header file A, there needs to be at
least a forward declaration of class A.

C++ doesn't care where this forward declaration is (or even if it
appears multiple times), but it seems better to do it this way in the
spirit of better "separation of concerns". It is the responsibility of
each header file to at least declare its class at all times, which it
can do as the first thing, since no dependencies are needed for a
forward declaration.