Translate Ast.expr.InternalName to private (not public) field refs
This change is a prerequisite to solve
https://github.com/kaitai-io/kaitai_struct/issues/1117. If the
implementation of the `valid` key calls the public accessors to get the
value to validate, modifying the `valid` implementation to apply the
`valid` constraint to individual elements (as required by
https://github.com/kaitai-io/kaitai_struct/issues/1117) causes infinite
recursion in the ValidFailInst test in those languages (C++/STL, C#, Go)
which use flags to remember whether an instance already has a cached
value or not. This is because the public accessor of the instance is
then called from the body of the public accessor itself (recursively),
and the recursion is infinite due to running the `valid` check before
setting the instance flag.