Lua: theoretical support for multi-byte terminators
See https://github.com/kaitai-io/kaitai_struct/issues/187

The new methods used in this commit have been implemented in
https://github.com/kaitai-io/kaitai_struct_lua_runtime/commit/aa9fa8445e80f27ea52b97135e90db8edd16c12b

Unfortunately, it cannot be said that this adds UTF-16 and UTF-32
support to `type: strz`, because the Lua runtime library doesn't support
these encodings yet. It only supports ASCII and UTF-8, see
https://github.com/kaitai-io/kaitai_struct_lua_runtime/blob/a7b8a9144f0978cd75d3b176c61a773be56370c0/string_decode.lua#L35-L43

So this is more of a promise of future support for
https://github.com/kaitai-io/kaitai_struct/issues/187 once UTF-16 and
UTF-32 encodings are implemented.