JavaScript: assign validation errors to item (not array) in `_debug`
This fixes incorrect display of validation errors for repeated fields in
the Web IDE. For example, if you load this pair of files in the Web IDE:

* [`valid_fail_repeat_contents.ksy`](https://github.com/kaitai-io/kaitai_struct_tests/blob/9f782819b70db6069674995b969d2b801c86a515/formats/valid_fail_repeat_contents.ksy)
* [`bcd_user_type_be.bin`](https://github.com/kaitai-io/kaitai_struct_tests/blob/9f782819b70db6069674995b969d2b801c86a515/src/bcd_user_type_be.bin)

... the object tree (when using a version of KSC from before this
commit) looked like this:

```
|
+-- foo  X
    |
    +-- 0 = [18, 52, 86, 120]
    +-- 1 = [18, 52, 86, 120]
    +-- 2 = [0, 18, 52, 86]
```

Note that the cross mark is shown next to the `foo` node as if the
entire array was parsed correctly but then failed to validate as a
whole. However, the cross mark is obviously misplaced: the failed
validation only involved the `foo[2]` entry, so the cross mark should be
there instead. This is fixed in this commit, so that the new object tree
in the same situation looks like this:

```
|
+-- foo /!\
    |
    +-- 0 = [18, 52, 86, 120]
    +-- 1 = [18, 52, 86, 120]
    +-- 2 = [0, 18, 52, 86]  X
```

Now it looks right, including the warning icon on the `foo` node,
indicating that this array might not have been fully parsed.