C++, C#, Go: unconditionaly set instance flags before evaluation
Fixes a C++98 memory leak in the ValidFailRepeatInst test.

This commit also fixes "broken caching" (leading to repeated evaluation
of the same instance if invoked again) in those target languages that
use instance flags, i.e. C++, C# and Go, if:

1. the first evaluation raised an exception (such as in the
ValidFailRepeatInst test format) or
2. the `if` expression specified for the instance evaluated to `false`
in the first evaluation.

   Previously, as seen in the generated code in C++ and C# from test
formats IfValues and IfInstances, the `f_* = true;` assignment was done
outside the `if` statement for value instances, but inside the `if`
statement for parse instances.

Unfortunately, this is not yet tested in KST (and I recall that the
languages that don't use instance flags had similar problems).