BaseTranslator: remove default doByteArrayLiteral() implementation
Now that I've fixed the translation of byte array literals in
`JavaScriptTranslator` in the previous commit, which was the only place
where this default implementation of `doByteArrayLiteral()` was actually
used without being overridden (and it turned out to be a bug - it should
have been overridden there as well), I think it's best to remove it and
require each language-specific translator to provide the correct
implementation. The fact that all languages override
`BaseTranslator.doByteArrayLiteral()` proves that it is not a useful
implementation. On the contrary, it seems rather harmful: if
`BaseTranslator` had never implemented `doByteArrayLiteral()`, whoever
implemented `JavaScriptTranslator` would have been forced to implement it
explicitly, and thus the bug described and fixed in the previous commit
probably wouldn't have occurred at all.